+++ .dotnet/src/Custom/Assistants/ThreadMessage.cs
@@ -1,7 +1,9 @@
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 
 namespace OpenAI.Assistants;
 
+[Experimental("OPENAI001")]
 [CodeGenModel("MessageObject")]
 public partial class ThreadMessage
 {