+++ .dotnet/src/Custom/Assistants/ToolDefinition.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Diagnostics.CodeAnalysis;
 
 namespace OpenAI.Assistants;
 
+[Experimental("OPENAI001")]
 [CodeGenModel("AssistantToolDefinition")]
 public abstract partial class ToolDefinition
 {